backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-SocketIO==5.3.5
marshmallow==3.20.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
google-generativeai==0.3.1
email-validator==2.1.0
Werkzeug==3.0.1
requests==2.31.0
redis==5.0.1
celery==5.3.4
python-socketio==5.10.0
gunicorn==21.2.0
